requirements.txt
langchain
langchain-google-genai
langchain-openai
requests
beautifulsoup4
python-dotenv
tavily-python